requirements.txt
# FastAPI Backend
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.9.2

# Streamlit Frontend
streamlit==1.39.0
streamlit-option-menu==0.3.13

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2

# AI/ML
openai==1.51.2
anthropic==0.34.2
sentence-transformers==3.2.0
chromadb==0.5.15

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0
loguru==0.7.2

# Testing
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0